Added the editor/ml/make_mp_version.ml program to *generate* the mp_version.ml file.  This works, and it may give us extra flexibility in printing the version string.  However, other solutions that work on win32 and Unix are welcome.
WARNING WARNING WARNING: the Weak module in weak_memo.ml is now enabled.
This hides the weak memo bug, so if you are working on it, be sure to
change the module name to something else.  However, let's leave it like
this in the meantime.  It makes running MetaPRL easier, especially on
win32.


git-svn-id: http://svn.metaprl.org/svnroot/mojave/libmojave@175 5109f7fa-8300-0410-942d-a7a480986ffa