For the formatter, added an "azone" that assumes that its body is "atomic."  That is, the body is to be typeset all or none.  The width of an atomic block is 1 unit.
This addresses the formatting problem where multi-character
sequences were getting split.  By default, if we use UTF-8, the
formatter would interpret a multi-character unicode sequence
as multiple characters.  By wrapping it:
    azone <utf-8-sequence> ezone
the utf-8-sequence gets interpreted as an atomic unit.

This should work properly in prl and html mode.  The TeX version
is not implemented, but is no big deal for the moment.

For PRL mode, we may or may not want to split multiple character
Unicode sequences, like atomic["������������������"] (Leftrightarrow)
since it is actually a sequence of several unicode characters.


git-svn-id: http://svn.metaprl.org/svnroot/mojave/libmojave@357 5109f7fa-8300-0410-942d-a7a480986ffa