Oops, the _ variable gets turned into an empty variable.  I'm not sure if we should fix Lm_symbol, but in any case, I've updated Filter_patt to consider "" to denote a wildcard pattern.
Also, when we use simple wildcards, the syntax become unreadable
(at least to me).  OCaml plans to treat variables that begin with an
underscore as real variables, but with special meaning wrt dead code.
In fact, it may already do this.

To handle these issues, Filter_patt now considers the following veriables
to denote wildcards:
   1. the empty string,
   2. a single underscore,
   3. a double underscore followed by anything.


git-svn-id: http://svn.metaprl.org/svnroot/mojave/libmojave@843 5109f7fa-8300-0410-942d-a7a480986ffa