Working on outstanding issues for Win32.
This is a fix to be more careful about spaces in filenames,
so that placing ocaml in

   C:\Program Files\Objective Caml

actually works.  This adds the $(quote-argv ...) function that
quotes an array in a way that the command-line parser can recover
the argv string.  This allows include directories with spaces in
a -pp option.


git-svn-id: http://svn.metaprl.org/svnroot/mojave/libmojave@921 5109f7fa-8300-0410-942d-a7a480986ffa